+++ phenum/io_utils.py
@@ -280,18 +280,23 @@ def read_enum_out(args):
 
     return (system, structure_data)
 
-def write_POSCAR(system_data,space_data,structure_data,displacement):
+def write_POSCAR(system_data,space_data,structure_data,args):
     """Writes a vasp POSCAR style file for the input structure and system
     data.
 
     :arg system_data: a dictionary of the system_data
     :arg space_data: a dictionary containing the spacial data
     :arg structure_data: a dictionary of the data for this structure
+    :arg args: Dictionary of user supplied input.
     """
 
     from numpy import array
-    
-    filename = "vasp.{}".format(str(structure_data["strN"]))
+    from element_data import get_lattice_parameter
+
+    if "{}" in args["outfile"]:
+        filename = args["outfile"].format(str(structure_data["strN"]))
+    else:
+        filename = args["outfile"] + ".{}".format(str(structure_data["strN"]))
 
     labeling = structure_data["labeling"]            
     gIndx = space_data["gIndx"]
@@ -299,28 +304,41 @@ def write_POSCAR(system_data,space_data,structure_data,displacement):
 
     arrow_directions = [[0,0,0],[0,0,-1],[0,-1,0],[-1,0,0],[1,0,0],[0,1,0],[0,0,1]]
     directions = []
+
+    concs = []
+    for i in range(system_data["k"]):
+        this_conc = 0
+        for atom in range(structure_data["n"]*system_data["nD"]):
+            if labeling[gIndx[atom]] == str(i):
+                this_conc += 1
+        concs.append(this_conc)
+    def_title = "{} str #: {}\n".format(str(system_data["title"]),str(structure_data["strN"]))
+
+    lattice_parameter, title = get_lattice_parameter(args["species"],concs,def_title)
     for arrow in arrows:
         directions.append(array(arrow_directions[int(arrow)]))
-    
+
     with open(filename,"w+") as poscar:
-        poscar.write("{} str #: {}\n".format(str(system_data["title"]),str(structure_data["strN"])))
-        poscar.write("1.00\n")
+        poscar.write("{}\n".format(title))
+        poscar.write("{}\n".format(lattice_parameter))
         for i in range(3):
             poscar.write(" {}\n".format(" ".join(
                 ["{0: .8f}".format(j) for j in space_data["sLV"][i]])))
         poscar.write(" ")
-        for i in range(system_data["k"]):
-            ic = 0
-            for iAt in range(structure_data["n"]*system_data["nD"]):
-                if labeling[gIndx[iAt]] == str(i):
-                    ic += 1
-            poscar.write("{}   ".format(str(ic)))
+        if args["species"] == None:
+            for ic in concs:
+                poscar.write("{}   ".format(str(ic)))
+        else:
+            for ic in concs:
+                if ic != 0:
+                    poscar.write("{}   ".format(str(ic)))                    
+
         poscar.write("\n")
         poscar.write("D\n")
         for ilab in range(system_data["k"]):
             for iAt in range(structure_data["n"]*system_data["nD"]):
                 if labeling[gIndx[iAt]] == str(ilab):
-                    out_array = array(space_data["aBas"][iAt]) + directions[iAt]*displacement
+                    out_array = array(space_data["aBas"][iAt]) + directions[iAt]*args["displace"]
                     poscar.write(" {}\n".format(
                         "  ".join(["{0: .8f}".format(i) for i in out_array.tolist()])))